+++ model/schema_combo_test.go
@@ -5,4 +5,5 @@ import "testing"
 func TestCombo(t *testing.T) {
 	// TODO: Schema.GetOrCreate
 	// TODO: Schema.GetOrCreateTx
+	// TODO: Schema.GetAndLockTx
 }